Trait ark_crypto_primitives::SNARK

pub trait SNARK<F> where
    F: PrimeField, {
    type ProvingKey: Clone;
    type VerifyingKey: Clone + ToBytes;
    type Proof: Clone;
    type ProcessedVerifyingKey: Clone;
    type Error: 'static + Error;
    fn circuit_specific_setup<C, R>(
        circuit: C, 
        rng: &mut R
    ) -> Result<(Self::ProvingKey, Self::VerifyingKey), Self::Error>
    where
        C: ConstraintSynthesizer<F>,
        R: RngCore + CryptoRng;

    fn prove<C, R>(
        circuit_pk: &Self::ProvingKey, 
        circuit: C, 
        rng: &mut R
    ) -> Result<Self::Proof, Self::Error>
    where
        C: ConstraintSynthesizer<F>,
        R: RngCore + CryptoRng;

    fn process_vk(
        circuit_vk: &Self::VerifyingKey
    ) -> Result<Self::ProcessedVerifyingKey, Self::Error>;

    fn verify_with_processed_vk(
        circuit_pvk: &Self::ProcessedVerifyingKey, 
        public_input: &[F], 
        proof: &Self::Proof
    ) -> Result<bool, Self::Error>;

    fn verify(
        circuit_vk: &Self::VerifyingKey, 
        public_input: &[F], 
        proof: &Self::Proof
    ) -> Result<bool, Self::Error> { ... }
}

The basic functionality for a SNARK.

Associated Types

type ProvingKey: Clone

The information required by the prover to produce a proof for a specific circuit C.

type VerifyingKey: Clone + ToBytes

The information required by the verifier to check a proof for a specific circuit C.

type Proof: Clone

The proof output by the prover.

type ProcessedVerifyingKey: Clone

This contains the verification key, but preprocessed to enable faster verification.

type Error: 'static + Error

Errors encountered during setup, proving, or verification.

Required methods

fn circuit_specific_setup<C, R>(
circuit: C,
rng: &mut R
) -> Result<(Self::ProvingKey, Self::VerifyingKey), Self::Error> where
C: ConstraintSynthesizer<F>,
R: RngCore + CryptoRng,

Takes in a description of a computation (specified in R1CS constraints), and samples proving and verification keys for that circuit.

fn prove<C, R>(
circuit_pk: &Self::ProvingKey,
circuit: C,
rng: &mut R
) -> Result<Self::Proof, Self::Error> where
C: ConstraintSynthesizer<F>,
R: RngCore + CryptoRng,

Generates a proof of satisfaction of the arithmetic circuit C (specified as R1CS constraints).

fn process_vk(
circuit_vk: &Self::VerifyingKey
) -> Result<Self::ProcessedVerifyingKey, Self::Error>

Preprocesses circuit_vk to enable faster verification.

fn verify_with_processed_vk(
circuit_pvk: &Self::ProcessedVerifyingKey,
public_input: &[F],
proof: &Self::Proof
) -> Result<bool, Self::Error>

Checks that proof is a valid proof of the satisfaction of circuit encoded in circuit_pvk, with respect to the public input public_input, specified as R1CS constraints.

Provided methods

fn verify(
circuit_vk: &Self::VerifyingKey,
public_input: &[F],
proof: &Self::Proof
) -> Result<bool, Self::Error>

Checks that proof is a valid proof of the satisfaction of circuit encoded in circuit_vk, with respect to the public input public_input, specified as R1CS constraints.